Moved the notifs item decoration to its own class for reuse in the order list
