+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/notifications/NotifsListFragment.kt
@@ -28,11 +28,11 @@ import com.woocommerce.android.tools.SelectedSite
 import com.woocommerce.android.ui.base.TopLevelFragment
 import com.woocommerce.android.ui.base.UIMessageResolver
 import com.woocommerce.android.ui.main.MainNavigationRouter
-import com.woocommerce.android.ui.notifications.NotifsListAdapter.ItemType
-import com.woocommerce.android.ui.notifications.NotifsListAdapter.NotifsListItemDecoration
 import com.woocommerce.android.util.WooLog
 import com.woocommerce.android.util.WooLog.T.NOTIFICATIONS
 import com.woocommerce.android.widgets.AppRatingDialog
+import com.woocommerce.android.widgets.BadgedItemDecoration
+import com.woocommerce.android.widgets.BadgedItemDecoration.ItemDecorationListener
 import com.woocommerce.android.widgets.SkeletonView
 import com.woocommerce.android.widgets.sectionedrecyclerview.SectionedRecyclerViewAdapter.Companion.INVALID_POSITION
 import dagger.android.support.AndroidSupportInjection
@@ -48,7 +48,7 @@ import javax.inject.Inject
 class NotifsListFragment : TopLevelFragment(),
         NotifsListContract.View,
         NotifsListAdapter.ReviewListListener,
-        NotifsListAdapter.ItemDecorationListener {
+        ItemDecorationListener {
     companion object {
         val TAG: String = NotifsListFragment::class.java.simpleName
         const val KEY_LIST_STATE = "list-state"
@@ -138,7 +138,7 @@ class NotifsListFragment : TopLevelFragment(),
 
         notifsAdapter.setListListener(this)
 
-        val unreadDecoration = NotifsListItemDecoration(activity as Context)
+        val unreadDecoration = BadgedItemDecoration(activity as Context)
         unreadDecoration.setListener(this)
 
         notifsList.apply {
@@ -452,7 +452,7 @@ class NotifsListFragment : TopLevelFragment(),
     /**
      * Determines whether to show the unread indicator item decoration for the passed position
      */
-    override fun getItemTypeAtPosition(position: Int): ItemType {
+    override fun getItemTypeAtPosition(position: Int): BadgedItemDecoration.ItemType {
         return notifsAdapter.getItemTypeAtRecyclerPosition(position)
     }
 }
